pyproject.toml
[project]
name = "tool-hitl-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=5.1.0",
    "click>=8.3.1",
    "fastapi>=0.128.5",
    "langchain-core>=1.2.9",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
]
